refactor (structure) 패키지 분리
