+++ src/main/java/step3/enums/Award.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.enums;
 
 public enum Award {
