+++ src/main/java/step3/enums/Message.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.enums;
 
 public enum Message {
