+++ src/test/java/step3/LottoTest.java
@@ -3,6 +3,7 @@ package step3;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Lotto;
 
 public class LottoTest {
